+++ src/renderer/src/App.tsx
@@ -7,6 +7,7 @@ import { DiagnosticsProvider } from './store/diagnostics'
 import { SettingsProvider } from './store/settings'
 import { ConsoleProvider } from './store/console'
 import { SyncProvider } from './store/sync'
+import { FileSelectionProvider } from './store/file-selection'
 import { LayoutProvider } from './store/layout'
 import { TutorialsProvider } from './store/tutorials'
 import { IS_WEB } from './lib/env'
@@ -19,19 +20,24 @@ function App(): JSX.Element {
             panel sizes map to the right slots (#528). */}
         <LayoutProvider chatPane={!IS_WEB}>
           <WorkspaceProvider>
-            <SyncProvider>
-              <DiagnosticsProvider>
-                <ConsoleProvider>
-                  <TutorialsProvider>
-                    <AppShell />
-                    <UpdateNotifier />
-                    {/* The refactoring diff preview (#634): renders nothing
+            {/* What is highlighted in each half of the Files panel (#848).
+                Above SyncProvider because folder sync needs the device
+                selection to know where a synced folder lands. */}
+            <FileSelectionProvider>
+              <SyncProvider>
+                <DiagnosticsProvider>
+                  <ConsoleProvider>
+                    <TutorialsProvider>
+                      <AppShell />
+                      <UpdateNotifier />
+                      {/* The refactoring diff preview (#634): renders nothing
                         until a refactoring is proposed from the editor. */}
-                    <RefactorPreview />
-                  </TutorialsProvider>
-                </ConsoleProvider>
-              </DiagnosticsProvider>
-            </SyncProvider>
+                      <RefactorPreview />
+                    </TutorialsProvider>
+                  </ConsoleProvider>
+                </DiagnosticsProvider>
+              </SyncProvider>
+            </FileSelectionProvider>
           </WorkspaceProvider>
         </LayoutProvider>
       </SettingsProvider>